Split off tests into separate files
