+++ tests/__init__.py
@@ -1,2 +1,11 @@
+from .appvar import *
+from .array import *
+from .file import *
+from .flash import *
+from .gdb import *
+from .group import *
+from .numeric import *
+from .picture import *
 from .pil import *
-from .tivars import *
+from .settings import *
+from .tokenization import *
